backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
groq
httpx